feat: implement BridgeInProgress component and related hooks for managing bridging steps and master safe creation
